fix: Attributes of one to many reflected references are mangled
In following scenario:

- Entity `category` refers `product` by reflected reference of cardinality zero or more and inherited attributes
- Entity `product` refers `category` by standard reference with attribute `variant` with cardinality zero or more

When product is created and contains two refrences:

1. category PK=1 and variant = `false`
2. category PK=2 and variant = `true`

The category with PK = 1 contains reflected reference to product with variant attribute set to `true` instead of false. It's because there are implicit attribute mutations by product for category with PK retrieved from all references to this entity type, so that there is:

1. upsertAttributeMutation: vairant = `false`
2. upsertAttributeMutation: vairant = `true`

And the latter overwrites the former. Correctly there should be only single attribute mutation that correctly targets category with PK = 1.

Refs: #803