feat: Accept labels in HTTP headers
Adding labels directly into the query might be problem in Node.js environment due to the use of query templates, but it would be quite easy to add them to the HTTP request as headers. So we should be able to optionally accept query labels in headers in the form of:

- `X-Forwarded-Uri`: as the original URI requested for the frontend facade
- `X-Meta-Label`: with arbitrary value for new labels passed in headers

These headers should be automatically accepted and incorporated in the traffic recording.

Refs: #802