fix: Closeable streams needs to be closed forcefully on session close
There are a few stateful streams that keep an open handle to a file. Although we state in the documentation that the reader is responsible for closing them, this is very often not done. Even in our tests there are only a few cases where we forgot to close these streams. However, if the session in which the stream is opened is closed, we should make sure that any such streams opened in that session are also closed.

Refs: #819