feat: Support for definition Sortable Attribute Compound in a Java annotation
Currently it's not possible to set up SAC by annotation in Java definition class as other schema components are. In this issue such annotation should be added. Because the SAC doesn't have any other sensible content other than the definition itself, it makes sense to apply such annotation on a class level and name set of attribute names and behaviour in it.

Refs: #842