fix: Session can be created during catalog replacement
Catalog renaming/replacing/closing does not effectively prevent a new session from being created if there is a lot of traffic on the catalogue. Method close all sessions should enable some kind of flag that will prevent new session to pop-up. This flag needs to be reseted when the catalogue leaves critical section and starts to be available again. Or it needs to be removed from the catalogue listing much earlier.

I've written an integration test that uses 100 parallel clients doing 1000 queries and when 50% queries are finished I make the catalog replacement. This test now passes with zero query failures.

Refs: #850