feat: Support for checking catalog state on EvitaContract level
Currently it's not possible to tell from the EvitaContract interface if the catalogue is corrupted or what its state is. Clients may then attempt to query or update such a catalogue without any chance of discovering that the catalogue is corrupt. Any such call will eventually end with an exception that they cannot avoid.

We should add a new method to this interface that would allow clients to determine the actual state of the catalogue.

Refs: #860