fix: corrected storage protocol upgrade mechanism
WAL migration must count with readTotal instead of record end, which may be skewed because not all data are read through ObservableInput.

Also we check that record is not larger than file size. Some files might have dangling - unused records at the end of the file, when the database is shutdown forcefully (by kill switch).