fix: Old WAL files might not contain statistic tail written
Under certain conditions (we are not yet sure which ones), the old WAL files, which are not written to, lack the tail statistics information in them. These statistics are used to locate transactions in them for the sake of the CDC, and also for replay in case evitaDB is restarted or transaction processing lags behind transaction commits.

The current handling is not secure enough and needs to be updated:

- when the catalogue is loaded, old WAL files need to be checked too (not just the current WAL file)
- if old WAL files lack tail information, they need to be checked, trimmed and the statistics appended to the tail
- all WAL files need to be checked to see if they contain consecutive monotonic rows of transactions
- Catalogue needs to be checked that the catalogue versions in: bootstrap record, catalogue header and WAL file match after all transactions from the WAL files have been applied.

This may reveal why / when the statistical information is missing from the old files. And if such a situation happens, it could be recovered automatically.

Refs: #873