feat!: Propagate catalog schema version to client
As for now the catalog schema on the client side is invalidate regurarly on obsolete checks and doesn't immediatelly detect catalog schema changes executed by different clients. We should return current catalog schema version at the close session method and similar places so that the client can immediatelly detect and invalidate the catalog schema.

BREAKING-CHANGE: `CompletableFuture` in `EvitaContract` and `EvitaSessionContract` were replaced with `CompletionStage`, which is read-only (intended) and can be easily converted to `CompletableFuture`.