fix: Collection is compacted even if no modification happens
There are some edge cases, such as restoring a backup on an Evita instance with a different compaction configuration, that lead to a situation in which, during the next transaction flush, the entity collection that is not the target of any modifications is compacted at the end of the flush operation.

This situation triggers a few additional bugs: the entity collection is not recreated, so the next catalog version uses the old persistence service, and compaction occurs with every transaction flush. Additionally, missing equals/hashcode implementations prevented the introduction of a new persistence service.

All the bugs need to be corrected, and compaction should never occur if the collection doesn't observe any modifications during a particular transaction flush.

Refs: #908