feat: GoLive needs to be changed to asynchronous operations
It could take considerable amount of time if it is combined with large session performing big data insertion. It needs to behave the same way as asynchronous close (because it does the goLiveAndClose - so it inherently closes sessions). At the end of the session the flush is performed which may wait for flushing all data to the disk.

Refs: #910