feat: Introduce more granular control over reduced indexes in reference schema
We realized that many secondary - ReducedEntityIndexes may not be necessary and propose to add better granularity to reference schema `indexed` property. Currently it could be only true / false and tied to particular scope. This setting then leads to creating ReducedEntityIndex for each of the reference. This might lead to explosion of indexed (in combination with reflected schemas).

In fact we need to distinguish two situations:

1. we need the indexed=true for ability to issue filters like `referenceHaving`
2. we need the indexed=true to shard the data of the original entity and gain a considerable performance benefit when applying filters/sort over the main entity

In many cases we just need the scenario 1. and this cannot be stated in the schema currently.

So the proposal is tu change indexed to enum of two possible values (similar to enums in unique property):

- FOR_FILTERING
- FOR_FILTERING_AND_PARTITIONING

The second will maintain the current behavior. The first will just maintain ReducedTypeIndex, but no ReducedIndexes for all possible references. This may bring considerable memory savings and speeding up indexing process.

Refs: #920