perf: InScope LIVE + ARCHIVE prevents prefetching
When query uses both scopes for filtering, the result formula tree is always wrapped in OR container, which prevents prefetching mode of the query processing. When client uses exact `entityPrimaryKeyInSet` it means that the entity is either in LIVE or in ARCHIVE scope, but it can never be in both at the same time. So the prefetching strategy is still valid approach in this situation.

Also there is another optimization possibility. We could prune the formula tree safely by applying following rules:

- when conjunctive formula (AND for example) contains EMPTY formula, it can be immediatelly pruned
- when OR formula contains only single sub-formula, it could be replaced by this constraint

Pruning of conjunctive formula needs to be handled separately depending of upper constraints in the formula tree:

- if they are conjunctive - it needs to be itself replaced with EmptyFormula
- if they are disjunctive - it could be removed entirely

Refs: #939
Signed-off-by: Jan Novotný <novotnaci@gmail.com>