feat: Support for multiple references to the same entity with different attribute sets
Clients encountered situations in which they needed to declare multiple items to the same entity within a single reference. For example, there is a reference to "media" with two references to media ID = 5. The reason is that the references are further distinguished by related attributes, for example:

a reference to ID = 5, attribute "category" = "motive," attribute "order" = 1
a reference to ID 5 with the category attribute set to "motive" and the order attribute set to 1
a reference to ID 6 with the category attribute set to "gallery" and the order attribute set to 2
In these cases, the user wants to filter by the category attribute and sort by the order attribute. This case is currently not supported by evitaDB.

Refs: #906
Signed-off-by: Jan Novotný <novotnaci@gmail.com>