fix: avoid dropped references accumulation
When reading references from the disk we remove all dropped references. When duplicate references were introduced, it leads to situation where many relations are dropped and new relations are added - in such case the storage part would bloat with dropped relations that are not needed anymore this mechanism ensures, that the fact that the reference was dropped is persisted, but this fact doesn't accumulate and with next rewrite of the storage part on the disk all previously dropped references are removed.