fix: inherited attributes doesn't require reference to be indexed
We validate only attributes that are not inherited from the original reference. When attributes are inherited, we don't control their indexing / faceting settings - they are controlled by the original reference schema.

But we still allow the reference to be non-indexed and then the index is not present on the reflected reference either. We enforce reference to be indexed only when there is new explicitly defined attribute that needs indexing.