fix: Query targetting multiple scopes might return incorrectly filtered referenced entities
When query targets both archive and live scope and reference is indexed only in one of the scopes and filtering is used for the indexed relation, the result will contain also reference that is not matched by the filter even it should.

Even if the filter related to LIVE scope contains attribute status equals to ACTIVE, the returned entity might contain reference to LIVE entity with status PRIVATE. It's because the archive scope in combination with non indexed reference in that scope adds all referenced entities to the result set effectivelly overriding the filtering result.

It should add only the references to entities that are present in ARCHIVE scope this way. But references to LIVE entities must all comply with the filter. Also when the main product will be in ARCHIVE scope, the filter should be ignored and all references both to ARCHIVE and LIVE scope should be part of the result.

Refs: #1048